refactor: update layout imports and enhance UI components
