--- src/components/layout/index.ts
@@ -1,8 +1,0 @@
-export { default as Container } from "./Container";
-export { default as Grid } from "./Grid";
-export { default as Header } from "./Header";
-export { default as Footer } from "./Footer";
-export { default as Sidebar } from "./Sidebar";
-export { default as MobileNav } from "./MobileNav";
-export { default as Layout } from "./Layout";
-export { default as Navigation } from "./Navigation";
