+++ src/main.tsx
@@ -1,6 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./themed-variables.css";
+import "./styles/layout.css";
+import "./styles/ui.css";
 import "./index.css";
 import App from "./App.tsx";
