+++ src/styles/ui.css
@@ -1,0 +1,117 @@
+/* UI Components */
+
+/* Button Component */
+.button {
+	display: inline-flex;
+	align-items: center;
+	justify-content: center;
+	padding: var(--space-sm) var(--space-md);
+	border-radius: var(--border-radius-md);
+	font-size: var(--font-size-md);
+	font-weight: var(--font-weight-medium);
+	cursor: pointer;
+	transition: all var(--transition-fast);
+	border: 1px solid transparent;
+
+	&:disabled {
+		opacity: 0.5;
+		cursor: not-allowed;
+	}
+}
+
+.button--primary {
+	background-color: var(--color-primary);
+	color: var(--color-text-light);
+
+	&:hover {
+		background-color: var(--color-primary-dark);
+	}
+}
+
+.button--secondary {
+	background-color: var(--color-background-alt);
+	color: var(--color-text);
+
+	&:hover {
+		background-color: var(--color-background-hover);
+	}
+}
+
+.button--outline {
+	background-color: transparent;
+	border-color: var(--color-border);
+	color: var(--color-text);
+
+	&:hover {
+		background-color: var(--color-background-alt);
+	}
+}
+
+.button--icon {
+	padding: var(--space-sm);
+	border-radius: 50%;
+	width: 40px;
+	height: 40px;
+}
+
+.button--small {
+	padding: var(--space-xs) var(--space-sm);
+	font-size: var(--font-size-sm);
+}
+
+.button--large {
+	padding: var(--space-md) var(--space-lg);
+	font-size: var(--font-size-lg);
+}
+
+/* Card Component */
+.card {
+	background-color: var(--color-background-card);
+	border-radius: var(--border-radius-lg);
+	padding: var(--space-lg);
+	box-shadow: var(--shadow-sm);
+	transition: transform var(--transition-fast),
+		box-shadow var(--transition-fast);
+
+	&:hover {
+		transform: translateY(-2px);
+		box-shadow: var(--shadow-md);
+	}
+}
+
+.card__header {
+	margin-bottom: var(--space-md);
+}
+
+.card__title {
+	font-size: var(--font-size-lg);
+	font-weight: var(--font-weight-bold);
+	color: var(--color-text);
+	margin: 0;
+}
+
+.card__content {
+	color: var(--color-text);
+}
+
+.card__footer {
+	margin-top: var(--space-md);
+	padding-top: var(--space-md);
+	border-top: 1px solid var(--color-border);
+}
+
+/* Responsive Styles */
+@media (max-width: 768px) {
+	.button {
+		width: 100%;
+	}
+
+	.button--icon {
+		width: 36px;
+		height: 36px;
+	}
+
+	.card {
+		padding: var(--space-md);
+	}
+}
